Update Kotlin and dependencies
Move replicator lifecycle extension to android-ktx
Ensure asKtxFlow() throws errors
Fix sample app build
