+++ app/build.gradle
@@ -1,14 +1,13 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
 
 android {
-    compileSdkVersion 31
+    compileSdk 33
 
     defaultConfig {
         applicationId "com.molo17.couchbase.lite"
-        minSdkVersion 22
-        targetSdkVersion 31
+        minSdk 22
+        targetSdk 33
         versionCode 1
         versionName "1.0"
 
@@ -42,23 +41,22 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation libraries.kotlin
     implementation libraries.coroutines
-    implementation "com.couchbase.lite:couchbase-lite-android:2.7.0"
+    implementation "com.couchbase.lite:couchbase-lite-android:3.0.5"
 
-    implementation project(":kotlin")
+    implementation project(":android-ktx")
 
-    implementation "androidx.appcompat:appcompat:1.1.0"
-    implementation "androidx.core:core-ktx:1.2.0"
-    implementation "androidx.activity:activity-ktx:1.1.0"
-    implementation "androidx.recyclerview:recyclerview:1.1.0"
-    implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.2.0"
-    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.2.0"
+    implementation "androidx.appcompat:appcompat:1.6.0"
+    implementation "androidx.core:core-ktx:1.9.0"
+    implementation "androidx.activity:activity-ktx:1.6.1"
+    implementation "androidx.recyclerview:recyclerview:1.2.1"
+    implementation "androidx.lifecycle:lifecycle-livedata-ktx:2.5.1"
+    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.1"
     implementation "androidx.lifecycle:lifecycle-extensions:2.2.0"
-    implementation "androidx.constraintlayout:constraintlayout:1.1.3"
-    implementation "com.github.bumptech.glide:glide:4.11.0"
+    implementation "androidx.constraintlayout:constraintlayout:2.1.4"
+    implementation "com.github.bumptech.glide:glide:4.14.2"
 
     testImplementation libraries.unitTests
-    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
 }
