+++ app/src/main/java/com/molo17/couchbase/lite/DependencyContainer.kt
@@ -25,6 +25,7 @@ import com.couchbase.lite.DatabaseConfiguration
 import com.couchbase.lite.Replicator
 import com.couchbase.lite.ReplicatorConfiguration
 import com.couchbase.lite.URLEndpoint
+import com.molo17.couchbase.lite.android.ktx.bindToLifecycle
 import com.molo17.couchbase.lite.data.CouchbaseHotelRepository
 import com.molo17.couchbase.lite.domain.HotelsRepository
 import java.net.URI
@@ -40,7 +41,7 @@ object DependencyContainer : ViewModelProvider.Factory {
 
     operator fun invoke(): () -> ViewModelProvider.Factory = { this }
 
-    override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
         return when (modelClass) {
             MainViewModel::class.java -> MainViewModel(hotelsRepository)
             else -> error("${modelClass.simpleName} not handled in DependencyContainer")
@@ -61,9 +62,10 @@ object DependencyContainer : ViewModelProvider.Factory {
 
     private val replicator by lazy {
         val url = URLEndpoint(URI.create(BuildConfig.REPLICATOR_URL))
-        val config = ReplicatorConfiguration(database, url).apply {
-            authenticator = BasicAuthenticator(BuildConfig.REPLICATOR_USER, BuildConfig.REPLICATOR_PWD)
-        }
+        val config = ReplicatorConfiguration(database, url)
+            .setAuthenticator(
+                BasicAuthenticator(BuildConfig.REPLICATOR_USER, BuildConfig.REPLICATOR_PWD.toCharArray())
+            )
         Replicator(config)
     }
 }
