+++ library/build.gradle
@@ -2,11 +2,9 @@ apply plugin: 'kotlin'
 apply plugin: 'maven-publish'
 
 dependencies {
-    def couchbaseLite = "com.couchbase.lite:couchbase-lite-java:3.0.0"
+    def couchbaseLite = "com.couchbase.lite:couchbase-lite-java:3.0.5"
 
-    implementation libraries.kotlin
     implementation libraries.coroutines
-    compileOnly libraries.lifecycle
     compileOnly couchbaseLite
 
     testImplementation libraries.unitTests
