+++ library/src/test/java/com/molo17/couchbase/lite/QueryExtensionsKtTest.kt
@@ -21,11 +21,11 @@ import com.couchbase.lite.Query
 import com.couchbase.lite.QueryChange
 import com.couchbase.lite.QueryChangeListener
 import com.couchbase.lite.ResultSet
-import com.nhaarman.mockitokotlin2.any
-import com.nhaarman.mockitokotlin2.doAnswer
-import com.nhaarman.mockitokotlin2.doReturn
-import com.nhaarman.mockitokotlin2.mock
-import com.nhaarman.mockitokotlin2.verify
+import org.mockito.kotlin.any
+import org.mockito.kotlin.doAnswer
+import org.mockito.kotlin.doReturn
+import org.mockito.kotlin.mock
+import org.mockito.kotlin.verify
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.first
